NV2A : Split up in the same way as OpenXbox. This makes differences between the two easier to spot.
