+++ src/devices/video/EmuNV2A_PRMDIO.cpp
@@ -1,0 +1,21 @@
+DEVICE_READ32(PRMDIO)
+{
+	DEVICE_READ32_SWITCH() {
+	default:
+		DEBUG_READ32_UNHANDLED(PRMDIO);
+		break;
+	}
+
+	DEVICE_READ32_END(PRMDIO);
+}
+
+DEVICE_WRITE32(PRMDIO)
+{
+	switch (addr) {
+	default:
+		DEBUG_WRITE32_UNHANDLED(PRMDIO);
+		break;
+	}
+
+	DEVICE_WRITE32_END(PRMDIO);
+}
