+++ src/devices/video/EmuNV2A_PRMVIO.cpp
@@ -1,0 +1,21 @@
+DEVICE_READ32(PRMVIO)
+{
+	DEVICE_READ32_SWITCH() {
+	default:
+		DEBUG_READ32_UNHANDLED(PRMVIO); // TODO : DEVICE_READ32_REG(prmvio);
+		break;
+	}
+
+	DEVICE_READ32_END(PRMVIO);
+}
+
+DEVICE_WRITE32(PRMVIO)
+{
+	switch (addr) {
+	default:
+		DEBUG_WRITE32_UNHANDLED(PRMVIO); // TODO : DEVICE_WRITE32_REG(prmvio);
+		break;
+	}
+
+	DEVICE_WRITE32_END(PRMVIO);
+}
