+++ src/devices/video/EmuNV2A_PSTRAPS.cpp
@@ -1,0 +1,23 @@
+DEVICE_READ32(PSTRAPS)
+{
+	DEVICE_READ32_SWITCH() {
+	default:
+		DEBUG_READ32_UNHANDLED(PSTRAPS);
+		break;
+	}
+
+	DEVICE_READ32_END(PSTRAPS);
+}
+
+DEVICE_WRITE32(PSTRAPS)
+{
+	switch (addr) {
+	default:
+		DEBUG_WRITE32_UNHANDLED(PSTRAPS);
+		break;
+	}
+
+	DEVICE_WRITE32_END(PSTRAPS);
+}
+
+
